backend/requirements.txt
argcomplete==1.9.3
boto3==1.10.14
botocore==1.13.14
certifi==2019.9.11
cfn-flip==1.2.2
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.9
django-cors-headers==3.1.1
django-rest-swagger==2.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
docutils==0.15.2
durationpy==0.5
future==0.16.0
hjson==3.0.1
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.3
lambda-packages==0.20.0
MarkupSafe==1.1.1
numpy==1.17.3
openapi-codec==1.3.2
pandas==0.25.3
placebo==0.9.0
PyJWT==1.7.1
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.2.1
simplejson==3.16.0
six==1.13.0
toml==0.10.0
tqdm==4.19.1
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.25.6
Werkzeug==0.16.0
wsgi-request-logger==0.4.6
